+++ doc/references/release-notes.rst
@@ -23,6 +23,10 @@ Upcoming Release
   The new modularized maps module allows for more flexibility and easier
   maintenance. 
 
+* New method `n.equals() <pypsa.Network.equals>` to compare two networks for equality. 
+  This is similar to the equality operator `==` but allows for more flexibility in the
+  comparison which is useful for testing and debugging.
+
 `v0.34.0 <https://github.com/PyPSA/PyPSA/releases/tag/v0.34.0>`__ (25th March 2025)
 =======================================================================================